requirements.local.txt
starlette~=0.45
uvicorn~=0.34.0
python-dotenv==1.0.1